c310: extract SpectralBrowserTab + FalseColorBandPicker to own file (#441 2.1)
Workspace.tsx 3851 LOC -> 3413 LOC (-438 LOC). Entry chunk 103.57 KB -> 93.27 KB (-10.3 KB).
28 tabs extracted + lazy. Cumulative since c262: -70%.
Removed newly-unused SpectralBrowser import from Workspace.tsx.
typecheck clean, 44/44 tests pass.
